+++ app/src/main/java/com/example/betterworld/utils/Actions.java
@@ -54,7 +54,7 @@ public class Actions {
     }
 
 
-    public static void goToDonationDetailActivity(Activity activity,float amount,String charityId) {
+    public static void goToDonationDetailActivity(Activity activity,double amount,String charityId) {
         Intent intent = new Intent(activity, DonationDetailActivity.class);
         intent.putExtra("DONATION_AMOUNT", amount);
         intent.putExtra("CHARITY_ID", charityId);
